Organizando funções com Reduce
Esta sendo realizado a otimização de codigo utilizando o Reduce. Foi criado a pasta reducers, onde o componente cycles.ts esta lidando com as funções de controle de Ciclos da aplicação
